requirements.txt
pandas==2.2.3
numpy==2.2.5
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.6.1
joblib==1.5.0
xgboost==3.0.1
lightgbm==4.6.0
catboost==1.2.8
sqlalchemy==2.0.40
sqlalchemy-utils==0.41.2
kafka-python==2.2.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
jupyterlab==4.4.2
requests==2.32.3
pytz==2025.2
python-dateutil==2.9.0.post0
setuptools==80.3.1
wheel
